Refactor methods in python and language package services
